+++ spec/support/exist_matcher.rb
@@ -1,0 +1,6 @@
+RSpec::Matchers.define :exist do 
+  match do |actual|
+    actual.exists?
+  end
+end
+